[Refactor] #30 BuildConfig 수정
